QuantPulse-Backend/requirements-py313.txt
# QuantPulse Backend Dependencies - Pinned for Render (Python 3.10)

# Core Infrastructure
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv>=1.0.1
pydantic>=2.4.1
pydantic[email]>=2.4.1
starlette>=0.27.0  # Required for SessionMiddleware

# Database & ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.3
python-multipart>=0.0.6
slowapi>=0.1.9  # Rate limiting for attack prevention
authlib>=1.3.0  # OAuth support for Google login
itsdangerous>=2.1.2  # Secure token generation

# HTTP Client and Caching
httpx>=0.27.0
cachetools>=5.3.0

# Data and Math
yfinance==0.2.51
pandas==2.2.3
# numpy>=1.23.5,<2.0.0  <-- MODIFIED for Python 3.13
numpy>=1.23.5
scikit-learn==1.5.2
hmmlearn==0.3.3
joblib>=1.4.0

# NLP (news sentiment)
textblob>=0.18.0

# AI and Deep Learning
tensorflow>=2.20.0
huggingface-hub>=0.23.0

# Agentic Framework
crewai>=1.0.0
crewai-tools>=1.0.0
langchain-core>=0.3.27
langchain-groq>=0.2.1
